update files to fit requirement formats, update readme for run-pipeline / add-text-embedding-feature / tee commands
