+++ tl/cli/get-phrase-matches.py
@@ -33,80 +33,10 @@ def add_arguments(parser):
     parser.add_argument('--filter', action='store', nargs='?', dest='filter_condition',
                         default="", help="If set to run filtering, which kind of the data should keep.")
 
-def remove_previous_match_res(df):
-    to_drop_cols = ["kg_id", "kg_labels", "method", "retrieval_score", "retrieval_score_normalized"]
-    df = df.drop(columns=to_drop_cols, errors="ignore").drop_duplicates()
-    return df
-
-def combine_result(input_df, output_df, filter_condition):
-    import operator
-    def get_operator(s):
-        if s == ">":
-            return operator.gt
-        elif s == "=":
-            return operator.eq
-        elif s == "<":
-            return operator.le
-    # load expression
-    expressions_parsed = []
-    temp = ""
-    left, right, comparator = None, None, []
-    for each in filter_condition:
-        if each == " ":
-            continue
-        if each in "|&":
-            right = float(temp)
-            if left is None or comparator is None or right is None:
-                raise tl.exceptions.TLException("Filter equation {} seems invalid!".format(filter_condition))
-            expressions_parsed.append((left, comparator, right))
-            expressions_parsed.append(each)
-        elif each in ">=<":
-            comparator.append(get_operator(each))
-            left = temp
-            temp = ""
-        else:
-            temp += each
-
-    right = float(temp)
-    if left is None or comparator is None or right is None:
-        raise tl.exceptions.TLException("Filter equation {} seems invalid!".format(filter_condition))
-    expressions_parsed.append((left, comparator, right))
-
-    # get expression result
-    final_res = None
-    for i, each_compare in enumerate(expressions_parsed):
-        if i % 2 == 0:
-            for each_comparator in each_compare[1]:
-                each_res = each_comparator(input_df[each_compare[0]].astype(float).fillna(0), each_compare[2])
-                if final_res is None:
-                    final_res = each_res
-                else:
-                    final_res = each_res & final_res
-            # final_res =  final_res & (input_df[each_compare[0]].astype(float), each_compare[2])
-        else:
-            if each_compare == "&":
-                final_res = each_res & final_res
-            else:
-                final_res = each_res | final_res
-    
-    input_df_filtered = input_df[final_res]
-
-    # mark down the data we still remained, those are not needed add
-    existed_groups = set()
-    for each in input_df_filtered.groupby(by=['column', 'row']):
-        existed_groups.add(each[0])
-    # add those data not exist
-    for each in output_df.groupby(by=['column', 'row']):
-        if each[0] not in existed_groups:
-            input_df_filtered = input_df_filtered.append(each[1])
-    # sort by column, row for human view
-    input_df_filtered = input_df_filtered.astype({"column": int, "row": int})
-    input_df_filtered = input_df_filtered.sort_values(by=['column', "row"])
-
-    return input_df_filtered
 
 def run(**kwargs):
     from tl.candidate_generation import phrase_query_candidates
+    from tl.utility.filter import Filter
     import pandas as pd
     try:
         df = pd.read_csv(kwargs['input_file'], dtype=object)
@@ -115,7 +45,7 @@ def run(**kwargs):
             need_filter = True
 
         if need_filter:
-            query_input_df = remove_previous_match_res(df)
+            query_input_df = Filter.remove_previous_match_res(df)
         else:
             query_input_df = df
 
@@ -123,8 +53,9 @@ def run(**kwargs):
                                                         es_user=kwargs['user'],
                                                         es_pass=kwargs['password'])
         odf = em.get_phrase_matches(kwargs['column'], properties=kwargs['properties'], size=kwargs['size'], df=query_input_df)
+        
         if need_filter:
-            odf = combine_result(df, odf, kwargs["filter_condition"])
+            odf = Filter.combine_result(df, odf, kwargs["filter_condition"])
 
         odf.to_csv(sys.stdout, index=False)
     except:
